controller/requirements.txt
kopf==0.27
kubernetes==11.0.0
pyyaml==5.3.1